Add initial error handling
