+++ src/app.html
@@ -9,13 +9,7 @@
       name="description"
       content="Display your YouTube subscriptions in a deck overview and create custom subscription groups."
     />
-    <script
-      type="text/javascript"
-      src="https://www.youtube.com/iframe_api"
-      defer
-      crossorigin="anonymous"
-      referrerpolicy="no-referrer"
-    ></script>
+    <script type="text/javascript" src="https://www.youtube.com/iframe_api" defer referrerpolicy="no-referrer"></script>
     %sveltekit.head%
   </head>
   <body class="w-full h-full" data-sveltekit-preload-data="hover">
