+++ src/routes/+layout.svelte
@@ -2,4 +2,8 @@
   import "../app.css";
 </script>
 
-<slot />
+<main class="w-full h-full overflow-x-auto x-scroll bg-neutral-700 text-white">
+  <div class="w-full h-full min-w-[22rem]">
+    <slot />
+  </div>
+</main>
